Move api to pkg folder
