+++ pkg/onmetal/delete_machine_test.go
@@ -18,7 +18,7 @@ import (
 	"fmt"
 
 	"github.com/gardener/machine-controller-manager/pkg/util/provider/driver"
-	"github.com/onmetal/machine-controller-manager-provider-onmetal/api/v1alpha1"
+	"github.com/onmetal/machine-controller-manager-provider-onmetal/pkg/api/v1alpha1"
 	"github.com/onmetal/machine-controller-manager-provider-onmetal/pkg/internal"
 	computev1alpha1 "github.com/onmetal/onmetal-api/api/compute/v1alpha1"
 	testutils "github.com/onmetal/onmetal-api/utils/testing"
